+++ src/settings/FeeFinesTable/CopyModal.js
@@ -34,24 +34,14 @@ class CopyForm extends React.Component {
   static propTypes = {
     owners: PropTypes.arrayOf(PropTypes.object),
     onClose: PropTypes.func,
+    initialize: PropTypes.func.isRequired,
     handleSubmit: PropTypes.func.isRequired,
     intl: intlShape.isRequired,
   };
 
   constructor(props) {
     super(props);
-    this.state = {
-      option: true,
-    };
-    this.handleOption = this.handleOption.bind(this);
-  }
-
-  handleOption(e) {
-    const value = (e.target.value === 'true');
-
-    this.setState({
-      option: value,
-    });
+    props.initialize({ option: 'true' });
   }
 
   render() {
@@ -73,15 +63,11 @@ class CopyForm extends React.Component {
                 label={<FormattedMessage id="ui-users.feefines.modal.yes" />}
                 id="yes"
                 value="true"
-                checked={this.state.option}
-                onChange={this.handleOption}
               />
               <RadioButton
                 label={<FormattedMessage id="ui-users.feefines.modal.no" />}
                 id="no"
                 value="false"
-                checked={!this.state.option}
-                onChange={this.handleOption}
               />
             </Field>
           </Col>